backend/requirements.txt
aiohttp==3.9.1
python-socketio==5.10.0
